Better status values, code cleanup, and file rename.
